relocate projectselect
